requirements.txt
mailshooter==0.0.1
Mako==1.0.3
MarkupSafe==0.23
PyYAML==3.11
wheel==0.26.0
